Move error file to match rest of commands
